pyproject.toml
[tool.poetry]
name = "psenet"
version = "0.1.0"
description = ""
authors = ["Sasha Illarionov <sasha@sdll.space>"]

[tool.poetry.dependencies]
python = "^3.7"
tensorflow = "^1.14"
segmentation-models = "^0.2.1"
opencv-python = "^4.1"
tqdm = "^4.32"
absl-py = "^0.7.1"
Polygon3 = "^3.0"
pyclipper = "^1.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = {version = "^3.7", allows-prereleases = true}
black = {version = "^18.3-alpha.0", allows-prereleases = true}

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

